+++ src/plugins/lua/weechat-lua-api.c
@@ -4001,6 +4001,24 @@ API_FUNC(buffer_get_integer)
     API_RETURN_INT(value);
 }
 
+API_FUNC(buffer_get_longlong)
+{
+    const char *buffer, *property;
+    long long value;
+
+    API_INIT_FUNC(1, "buffer_get_integer", API_RETURN_LONGLONG(-1));
+    if (lua_gettop (L) < 2)
+        API_WRONG_ARGS(API_RETURN_LONGLONG(-1));
+
+    buffer = lua_tostring (L, -2);
+    property = lua_tostring (L, -1);
+
+    value = weechat_buffer_get_longlong (API_STR2PTR(buffer),
+                                         property);
+
+    API_RETURN_LONGLONG(value);
+}
+
 API_FUNC(buffer_get_string)
 {
     const char *buffer, *property, *result;
@@ -5990,6 +6008,7 @@ const struct luaL_Reg weechat_lua_api_funcs[] = {
     API_DEF_FUNC(buffer_merge),
     API_DEF_FUNC(buffer_unmerge),
     API_DEF_FUNC(buffer_get_integer),
+    API_DEF_FUNC(buffer_get_longlong),
     API_DEF_FUNC(buffer_get_string),
     API_DEF_FUNC(buffer_get_pointer),
     API_DEF_FUNC(buffer_set),