+++ src/plugins/php/weechat-php_arginfo.h
@@ -1,5 +1,5 @@
 /* This is a generated file, edit weechat-php.stub.php instead.
- * Stub hash: 77ccd929a546ea5cebfdc39c333995e5b3fe3640 */
+ * Stub hash: 1bf8013bdcd61ae7e3193c29d5a2c4fabae8eadd */
 
 ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_weechat_register, 0, 7, IS_LONG, 0)
 	ZEND_ARG_TYPE_INFO(0, p0, IS_STRING, 0)
@@ -433,6 +433,8 @@ ZEND_END_ARG_INFO()
 #define arginfo_weechat_buffer_unmerge arginfo_weechat_mkdir_home
 
 #define arginfo_weechat_buffer_get_integer arginfo_weechat_string_has_highlight
+
+#define arginfo_weechat_buffer_get_longlong arginfo_weechat_string_has_highlight
 
 #define arginfo_weechat_buffer_get_string arginfo_weechat_iconv_to_internal